Tensor(s) for drawing batches from a dataset

Tensor or list(s) of tensors (e.g. for features and response) that yield the next batch of data each time they are evaluated.

batch_from_dataset(dataset, features = NULL, response = NULL,
  named = TRUE, named_features = FALSE)

Arguments

dataset
A dataset

features
Features to include.

response
Response variable.

named
TRUE to assign names to list elements when features are specified (defaults to x and y, provide a character vector to use alterate names).

named_features
TRUE to yield features as a named list; FALSE to stack features into a single array. Note that in the case of FALSE (the default) all features will be stacked into a single 2D tensor so need to have the same underlying data type.

Value

Tensor(s) that can be evaluated to yield the next batch of training data.

When features is specified the tensors will have a structure of either:

list(x = list(feature_name = feature_values, ...), y = response_values) when named_features is TRUE; or

list(x = features_array, y = response_values) when named_features is FALSE, where features_array is a Rank 2 array of (batch_size, num_features).

Note that by default list elements are named x and y. This can be customized using the named argument (pass FALSE to return a list with no names).

When features is not specified the tensors will conform to the shape and types of the dataset (see output_shapes() and output_types()).

Batch Iteration

In many cases you won't need to explicitly iterate over the batches, rather, you will pass the tensors to another function that will perform the evaluation (e.g. the Keras layer_input() and compile() functions).

If you do need to perform iteration manually by evaluating the tensors, there are a couple of possible approaches to controlling/detecting when iteration should end.

One approach is to create a dataset that yields batches infinitely (traversing the dataset multiple times with different batches randomly drawn). In this case you'd use another mechanism like a global step counter or check for a learning plateau.

Another approach is to iteration is to detect when all batches have been yielded from the dataset. When the batch tensor reaches the end of iteration a runtime error will occur. You can catch and ignore the error when it occurs by wrapping your iteration code in the with_dataset_iterator() function.

See the examples below for a demonstration of each of these methods of iteration.

See also

input_fn_from_dataset() for use with tfestimators.

Examples

# NOT RUN {
# iteration with 'infinite' dataset and explicit step counter

library(tfdatasets)
dataset <- csv_dataset("training.csv") %>%
  dataset_shuffle(5000) %>%
  dataset_batch(128) %>%
  dataset_repeat()
batch <- batch_from_dataset(dataset, features = c(mpg, disp), response = cyl)
steps <- 200
for (i in 1:steps) {
  # use batch$x and batch$y tensors
}

# iteration that detects and ignores end of iteration error

library(tfdatasets)
dataset <- csv_dataset("training.csv") %>%
  dataset_batch(128) %>%
  dataset_repeat(10)
batch <- batch_from_dataset(dataset, features = c(mpg, disp), response = cyl)
with_dataset_iterator({
  while(TRUE) {
    # use batch$x and batch$y tensors
  }
})
# }