Transform a dataset with delimted text lines into a dataset with named columns

Transform a dataset with delimted text lines into a dataset with named columns

dataset_decode_delim(dataset, record_spec, parallel_records = NULL)

Arguments

dataset
Dataset containing delimited text lines (e.g. a CSV)

record_spec
Specification of column names and types (see delim_record_spec()).

parallel_records
(Optional) An integer, representing the number of records to decode in parallel. If not specified, records will be processed sequentially.

See also

Other dataset methods: dataset_batch, dataset_cache, dataset_concatenate, dataset_filter, dataset_interleave, dataset_map, dataset_padded_batch, dataset_prefetch, dataset_repeat, dataset_shuffle, dataset_skip, dataset_take