Randomly shuffles the elements of this dataset.

Randomly shuffles the elements of this dataset.

dataset_shuffle(dataset, buffer_size, seed = NULL)

Arguments

dataset
A dataset

buffer_size
An integer, representing the number of elements from this dataset from which the new dataset will sample.

seed
(Optional) An integer, representing the random seed that will be used to create the distribution.

Value

A dataset

See also

Other dataset methods: dataset_batch, dataset_cache, dataset_concatenate, dataset_decode_csv, dataset_map, dataset_repeat, dataset_skip, dataset_take