Construct a tfestimators input function from a dataset

Construct a tfestimators input function from a dataset

input_fn_from_dataset(dataset, features, response = NULL)

Arguments

dataset
A dataset

features
The names of feature variables to be used.

response
The name of the response variable.

Value

An input_fn suitable for use with tfestimators train, evaluate, and predict methods

Details

Creating an input_fn from a dataset requires that the dataset consist of a set of named output tensors (e.g. like the dataset produced by the csv_dataset() function).