Tensor(s) for drawing batches from a dataset

Tensor or list(s) of tensors (e.g. for features and response) that yield the next batch of data each time they are evaluated.

batch_from_dataset(dataset, features = NULL, response = NULL,
  named = TRUE, named_features = FALSE)

Arguments

dataset
A dataset

features
Features to include.

response
Response variable.

named
TRUE to assign names to list elements when features are specified (defaults to x and y, provide a character vector to use alterate names).

named_features
TRUE to yield features as a named list; FALSE to stack features into a single array. Note that in the case of FALSE (the default) all features will be stacked into a single 2D tensor so need to have the same underlying data type.

Value

Tensor(s) that can be evaluated to yield the next batch of training data.

When features is specified the tensors will have a structure of either:

list(x = list(feature_name = feature_values, ...), y = response_values) when named_features is TRUE; or

list(x = features_array, y = response_values) when named_features is FALSE, where features_array is a Rank 2 array of (batch_size, num_features).

Note that by default list elements are named x and y. This can be customized using the named argument (pass FALSE to return a list with no names).

When features is not specified the tensors will conform to the shape and types of the dataset (see output_shapes() and output_types()).

Batch Iteration

In many cases you won't need to explicitly iterate over the batches, rather, you will pass the tensors to another function that will perform the evaluation (e.g. the Keras layer_input() and compile() functions).

If you do need to perform iteration manually by evaluating the tensors, a runtime error will occur when the iterator has exhausted all available elements. You can use the out_of_range_error() function to distinguish this error from other errors which may have occurred. For example:

library(tfdatasets)
dataset <- csv_dataset("training.csv") 
  dataset_batch(128) 
  dataset_repeat(10)
batch <- batch_from_dataset(dataset, features = c(mpg, disp), response = cyl))
tryCatch({
  while(TRUE) {
    # use batch$x and batch$y tensors
  }
},
error = function(e) {
  if (!out_of_range_error())
    stop(e)
})

See also

input_fn_from_dataset() for use with tfestimators.