Transform a dataset with delimted text lines into a dataset with named columns

Transform a dataset with delimted text lines into a dataset with named columns

dataset_decode_delim(dataset, delim = ",", col_names = NULL,
  col_types = NULL, col_defaults = NULL, num_parallel_calls = NULL)

Arguments

dataset
Dataset containing delimited text lines (e.g. a CSV)

delim
Character delimiter to separate fields in a record (defaults to ",")

col_names
Character vector with column names (or NULL to automatically detect the column names from the first row of the input file).

If col_names is a character vector, the values will be used as the names of the columns, and the first row of the input will be read into the first row of the datset. Note that if the underlying text file also includes column names in it's first row, this row should be skipped explicitly with dataset_skip().

If NULL, the first row of the input will be used as the column names, and will not be included in dataset.

col_types
Column types. If NULL, all column types will be imputed from the first 1000 rows on the input. This is convenient (and fast), but not robust. If the imputation fails, you'll need to supply the correct types yourself.

Types can be explicitliy specified in a character vector as "integer", "double", and "character" (e.g. col_types = c("double", "double", "integer").

Alternatively, you can use a compact string representation where each character represents one column: c = character, i = integer, d = double (e.g. col_types =ddi`).

col_defaults
List of default values which are used when data is missing from a record (e.g. list(0, 0, 0L). If NULL then defaults will be automatically provided based on col_types (0 for numeric columns and "" for character columns).

num_parallel_calls
(Optional) An integer, representing the number of elements to process in parallel If not specified, elements will be processed sequentially.

Note

The csv_dataset() function is a convenience wrapper for the text_line_dataset() and dataset_decode_delim() functions.

See also

Other dataset methods: dataset_batch, dataset_cache, dataset_concatenate, dataset_map, dataset_repeat, dataset_shuffle, dataset_skip, dataset_take